Fix availability_overrides FK: match tbl_users.id type
Change chef_id from BIGINT UNSIGNED to INT UNSIGNED to match the
tbl_users.id column type for foreign key compatibility.

+++ backend/database/migrations/2025_12_04_000001_create_availability_overrides_raw_sql.php
@@ -27,7 +27,7 @@ class CreateAvailabilityOverridesRawSql extends Migration
         DB::statement("
             CREATE TABLE tbl_availability_overrides (
                 id BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
-                chef_id BIGINT UNSIGNED NOT NULL COMMENT 'FK to tbl_users.id',
+                chef_id INT UNSIGNED NOT NULL COMMENT 'FK to tbl_users.id',
                 override_date DATE NOT NULL COMMENT 'Which date this override applies to',
                 start_time TIME NULL COMMENT 'Override start time, NULL = cancelled',
                 end_time TIME NULL COMMENT 'Override end time, NULL = cancelled',
